+++ lcheapo/lcfix.py
@@ -7,9 +7,6 @@ from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 from future.builtins import *  # NOQA @UnusedWildImport
 
-import sdpchain
-# Should actually be .lcheapo, but doesn't work if lcdump.py is called directly
-from lcheapo import (LCDataBlock, LCDiskHeader, LCDirEntry)
 import argparse
 import queue
 import os
@@ -19,81 +16,16 @@ import logging      # for logging information
 import datetime
 import sys
 
+from .lcheapo import (LCDataBlock, LCDiskHeader, LCDirEntry)
+from . import sdpchain
+
 # ------------------------------------
 # Global Variable Declarations
 # ------------------------------------
-versionString = "0.66"
-VERSIONS = """
-  VERSIONS
-     0.2:
-      - Handles spliced datafiles with zero-filled gaps:
-        + Adds dates to headers in gap
-        + Verifies that gap is the right size
-          (no time tear afterwards)
-      - Handles incomplete datafiles
-        + Warns the user that the data do not go as far as the
-          directory claims
-        + Changes the "directoryEntries" value in the header
-    0.3:
-      - Output starts with block #
-      - Output is streamlined for zero- filled gaps
-      - Confirm channel #: consecutive and < 4)
-      - Changed block# reported to conform to lcdump.py (tell()/512-1)
-      - Expanded BUG2 to handle 2 consecutive bad times/channel
-        (called BUG #2b)
-      - Time tears (MUST BE FIXED) are saved to timeTears.txt
-      - Bug #1s repeating every 500 blocks are identified as Bug #3
-        (can reduce the text output by a lot!)
-    0.4:
-      - Expanded BUG #2 to handle up to 3 consecutive bad times/channel
-        (BUG #2c)
-    0.45:
-      - Add --dryrun option (don't output fixed LCHEAPO file)
-      - Add check of non-time header values
-    0.46:
-      - By default, program names output file
-    0.5:
-      - Creates a JSON file with execution information
-      - Handles multiple lcheapo input files (should be different fragments of
-        same station, still only write one JSON file)
-    0.51:
-      - Names the JSON file based on the input filename (each input file
-        gets its own JSON)
-      - Add -F (forceTimes) option: to force time tags to be consecutive (only
-        use if all time tears are proven wrong)
-    0.52:
-      - The JSON file is ALWAYS named process-steps.json
-      - If there is already a process-steps.json, new information is appended
-    0.6:
-      - If multiple files specified, assume all are sections of same instrument
-        file, add header from first file to all others, output names include
-        timestamp of start of data
-      - If there is already a process-steps.json, new information is appended
-    0.61:
-      - Corrected process-steps.json to have steps as list, not dictionary
-    0.62:
-      - Fixed bug with "warnings" variable
-    0.63:
-      - Fixed bug making last directory entry for originally headerless files
-    0.64:
-      - Updated process-steps.json file to match process-steps.schema.json
-    0.65:
-      - Shifts to "force time stamp" for last 3 samples: avoids reading
-        beyond EOF
-    0.66:
-      - Added '-d', '-i' and '-o' options to match SDPCHAIN programs
-    0.67:
-      - Recognize '*.header.*' files as header+directory, without data
-      - DOESN"T WORK CORRECTLY!! CAT HEADER TO FIRST DATA!!!!
-    NOT YET:
-      - Force non-time header values
-      - Test to make sure first file has header, and subsequent don't
-        (requires new routine isHeader() in lcheapo.py)
-      - If first file doesn't have header, allow header creation
-        (use lcheader.py)
-      - Change directory entry creation to create a new one if original header
-        didn't have enough directory entries
-"""
+version = {}
+with open(os.path.join(os.path.dirname(__file__), "version.py")) as fp:
+    exec(fp.read(), version)
+
 warnings = 0  # count # of warnings
 
 
@@ -205,7 +137,7 @@ def main():
         sdpchain.make_process_steps_file(
             'lcfix',
             'Fix common bugs in LCHEAPO data files',
-            versionString,
+            version['__version__'],
             " ".join(sys.argv),
             startTimeStr,
             returnCode,
@@ -255,7 +187,7 @@ def getOptions():
     parser.add_argument("infiles", metavar="inFileName", nargs='+',
                         help="Input filename(s)")
     parser.add_argument("--version", action='version',
-                        version='%(prog)s {:s}'.format(versionString))
+                        version='%(prog)s {:s}'.format(version['__version__']))
     parser.add_argument("-v", "--verbose", dest="verbosity", default=0,
                         action="count",
                         help="Be verbose (-v = kind of, -vv = very)")
@@ -706,7 +638,7 @@ def __processInputFile(ifp1, fname, outFileRoot, lcHeader,
                                         lcData.changeTime(t)
                                     else:
                                         # Time tear (do not fix it!)
-                                        fmt = "{:8d}: Time Tear in Data. " +\
+                                        fmt = "{:8d}: Time Tear in Data.   " +\
                                               "CH{:d} Expected Time: {}, " +\
                                               "Got: {}"
                                         txt = fmt.format(currBlock,